Generated Property
Included a "Generated Property" attribute. If a SharePointDomainModel's
property is marked as "GeneratedProperty" then the Repository will not
try to set the value when updating or creating the item. It will still
be populated when getting items. This can be used when Workflows are
populating the columns.
Get by query has been edited to only retrieve the properties that are
specified in the domain model. This reduces the number of columns
returned by queries by about 30.