Allow download of Document
This addition allows you to mark a Domain Model as a "Document Domain
Model" which will allow you to call the Download method on the
SharePoint repository, pass in a FileStream to save the document to
disk.
A SharePointDocumentDomainModel can be used for any Document list,
rather than a custom list.